renamed navbar components to more closely match existing website
