+++ src/App.js
@@ -1,13 +1,11 @@
 import React from "react";
 import "./App.css";
 import { Route, Switch } from "react-router-dom";
+import Navbar from "./features/nav/Navbar";
 import Home from "./features/home/Home";
 import About from "./features/about/About";
-import Blog from "./features/blogs/Blog";
-import News from "./features/news/News";
-import Partners from "./features/partners/Partners";
-import Team from "./features/team/Team.js";
-import Navbar from "./features/nav/Navbar";
+import Blog from "./features/blog/Blog";
+import Donate from "./features/donate/Donate";
 import Contact from "./features/contact/Contact";
 
 function App() {
@@ -17,10 +15,8 @@ function App() {
       <Switch>
         <Route exact path="/" render={() => <Home />} />
         <Route exact path="/about" render={() => <About />} />
-        <Route exact path="/blogs" render={() => <Blog />} />
-        <Route exact path="/news" render={() => <News />} />
-        <Route exact path="/partners" render={() => <Partners />} />
-        <Route exact path="/team" render={() => <Team />} />
+        <Route exact path="/blog" render={() => <Blog />} />
+        <Route exact path="/donate" render={() => <Donate />} />
         <Route exact path="/contact" render={() => <Contact />} />
       </Switch>
     </div>
